requirements.txt
google-generativeai
Pillow
PyMuPDF
opencv-python
tesseract
numpy
tenacity